Rename mock files to be unique so pytest can import them
