+++ lambdas/setup_db/requirements.dev.txt
@@ -1,0 +1,8 @@
+-i https://pypi.org/simple
+atomicwrites==1.1.5
+attrs==18.1.0
+more-itertools==4.2.0
+pluggy==0.6.0
+py==1.5.3
+pytest==3.6.2
+six==1.11.0
